Refactor out classes & constants from root file
