+++ bin/user/weatherlink_live/static/version.py
@@ -1,0 +1,2 @@
+DRIVER_NAME = "WeatherLinkLive"
+DRIVER_VERSION = "1.1.0-b.1"
